requirements.txt
beautifulsoup4==4.12.3
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
chromadb==0.4.24
openai==1.23.2
python-dotenv==1.0.1
